web_server/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
google-cloud-storage>=2.14.0
